Moa.jsx
Updated version of Moa, combined with the create moa.

Create_moa.jsx is deleted
